comment from this commit:
nobody should be writing to closed sockets, but caller code can do two
writes within an event, have the first fail and disconnect the other
side (whose destructor then closes the calling object, but it's still
in a method), and then the now-dead object does its second write.
that is this case.  we just lie and say it worked.  it'll be dead soon
and won't be hurt by this lie.