Attempt to fix an issue with Perlbal and FotoBilder interaction causing undesirable behavior (premature termination of client connection).
Cause:

In the client mode upload to FotoBilder (Upload.pm), if the picture you
are uploading already exists on the server, the server will respond and
say that the picture has been uploaded.  This response is sent before any
data is actually read.  Subsequently, the BackendHTTP within Perlbal that
was handling this request will get the response, give it to the ClientProxy
object, and then close the client when it's done writing.

According to RFC2616, servers "SHOULD NOT" close the transport connection
until the client has finished sending the request.  The above behavior
flat out contradicts the strongly worded recommendation of the RFC.  The
RFC goes on to state that some clients interpret this prematurely closed
connection as a server error and they never read the response.  This seems
to be the case with the current FotoBilder Perl client as well as other
language's implementations of HTTP as well.

Solution:

This commit introduces the concept of a ClientProxy knowing how much data
it is supposed to be reading as well as whether it has gotten a response
from the backend.  It also removes the responsibility of closing the
underlying socket from the BackendHTTP and grants it to itself.

Now, when a request comes in, the ClientProxy notes how much body data
there should be (Content-Length rules as used in BackendHTTP--these were
moved up into HTTPHeaders as an instance method content_length).  Then,
when the BackendHTTP writes data to the client, a flag is toggled and if
there's no data left to read, the client closes.  If data is left, the
connection is held open and incoming data is simply discarded until we
reach the end of the stream.  At that point, since we have already sent the
response from the backend, we can simply close.

Standard timeouts still apply, so if no data is received for 30 seconds,
the ClientProxy object will still be closed.