Fix a bug in Perlbal.  This took hours and hours to track down and the solution isn't even the best.
Basically, open filehandles for PUT requests are created as an IO::Handle
object.  When we're done with the object and want to close it, we pass it
off to aio_close and tell it to handle it.

aio_close gets the file descriptor number out of the filehandle and calls
close on it directly.  It then passes control to the callback.

Before the callback actually gets called, as this is a multiple process
system to do the asyncronous work, a new connection comes in and is
allocated the file descriptor number that we just closed, since it's no
longer in use.  That's all well and good.

Until we get back to the callback and undef the filehandle, which then
causes Perl to execute a close on the file descriptor, thereby closing the
descriptor it no longer owned.

Ideal solution?  Marked as a FIXME for now.