Add new Perlbal functionality: directory index files.
On web_server services, you can now set "index_files" to be a comma
separated list of filenames.  When a user navigates to a directory, this
list is used to try to find an index file to display to the user.

This can be used with or without the "dirindexing" option.  If dirindexing
is on, then if no index file is found, a directory index is created and
served.  Else, a 200 message saying that directory listing is disabled is
sent out.