Add on-demand object tracking.  When an object is created, it basically records a backtrace of where this object came from, and stores it until the object is destroyed.
This can be turned on in the config file/management interface and is off by
default:

SERVER track_obj = 0 or 1

You can then see the currently tracked objects:

tracking

It sorts the objects by age and lists when they were created and the
backtrace leading up to their creation.  This will enable the tracking
of where exactly objects are created in order to find out which type of
BackendHTTP is getting leaked, etc.

Note that for this to work, DEBUG_OBJ must be on.  So you incur no penalty
if DEBUG_OBJ is off in a production environment.