add tests for verify_backend,
and fix a bug that was uncovered by yesterday's assertion of parameter
validation assertions in HTTPHeaders {res/req}_keep_alive.  the bug
was that after an OPTIONS request, the BackendHTTP's req_headers was
still empty (since we never formally turned the OPTIONS * string into
a request object, just sent it) and the updated {res/req}_keep_alive
correctly expected non-undef values.  so now i add a new flag to say
that it's an options response.