verify_backend_path configuration option
From: Adam Jacob
Subject: 	Re: Perlbal with Mongrels
Date: 	November 1, 2007 4:38:33 PM PDT

I sent a patch to allow this to be a bit more tuneable, which allows
this to work with Rails.  (If anyone can figure out how to get Rails
to respond to an OPTION * request, you are my hero.)  It apparently
got lost in the shuffle for the last release, so here it is, updated
for the current trunk (should apply directly to 1.60 as well)

It allows you to set a verify_backend_path, which defaults to the
current '*'.  For rails, you can set it to /, or any other url you
would like to route your OPTIONs request to.