Prevent a complex crash scenario.
I have an unautomated test that gets perlbal to crash right now, will make an automated test out of it tomorrow.

If perlbal has a service selector, performs a reproxy, and throws an error in the reproxy setup routine that gets shown to the user perlbal will crash because ->next_request on the backend doesn't get called until after the reproxy setup. Basically since the backend has made a successful response to us we should detach it from our client earlier, so that errors in the client can't cause us to end up in a mixed state.

This is the proper fix for the duck-typing fix I made in a branch a few weeks ago.